refactor: ValueObject's Column annotation location change it should be in to ValueObject not Entity
