+++ src/main/java/kitchenpos/apply/menus/tobe/domain/MenuGroup.java
@@ -1,9 +1,6 @@
 package kitchenpos.apply.menus.tobe.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.util.UUID;
 
 @Table(name = "menu_group")
@@ -13,7 +10,7 @@ public class MenuGroup {
     @Id
     private UUID id;
 
-    @Column(name = "name", nullable = false)
+    @Embedded
     private MenuGroupName name;
 
     protected MenuGroup() { }
