+++ src/main/java/kitchenpos/apply/menus/tobe/domain/MenuGroupName.java
@@ -1,12 +1,14 @@
 package kitchenpos.apply.menus.tobe.domain;
 
-import kitchenpos.support.ValueObject;
+import kitchenpos.support.domain.ValueObject;
 import org.springframework.util.StringUtils;
 
+import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class MenuGroupName extends ValueObject {
+    @Column(name = "name", nullable = false)
     private String value;
 
     protected MenuGroupName() { }
