+++ src/main/java/kitchenpos/apply/products/tobe/domain/ProductName.java
@@ -1,11 +1,13 @@
 package kitchenpos.apply.products.tobe.domain;
 
-import kitchenpos.support.ValueObject;
+import kitchenpos.support.domain.ValueObject;
 
+import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class ProductName extends ValueObject {
+    @Column(name = "name", nullable = false)
     private String value;
 
     protected ProductName() { }
